Allow multiple dots for admin email address
This is just a simple regex to match generic email addresses and doesn't
try to be too clever, but it is the case that we might see email
addresses at subdomains, so we should allow them. This fix loosens the
requirements on the admin email address and adds an associated unit
test.

Fixes #2